requirements.txt
matplotlib==3.7.0
numpy==1.25.0
numpy==1.23.5
opencv_python==4.7.0.72
Pillow==9.4.0
Pillow==9.5.0
streamlit==1.23.1
streamlit_lottie==0.0.5
